Make horrible hack to get the C juices flowing again
This is so very gross but it's late, I'm about ready for bed and
the main purpose of this evening's endeavour was to see a timestamp
in the program output window.

Also good to thaw out the C muscle memory, after not having cut much C
in the past 2 years, working in a python-centric full-time job.